bosser ensemble

+++ cub3d_struct.h
@@ -20,9 +20,9 @@ typedef struct s_res
 
 typedef struct s_color
 {
-	float r;
-	float g;
-	float b;
+	int r;
+	int g;
+	int b;
 }				t_color;
 
 typedef struct s_vec
